Refactor BuildOSC function for improved readability and maintainability
Summary of Changes:
1. Broke down the BuildOSC function into smaller, more manageable functions.
2. Created five new functions: AddStatusMessage, AddWindowActivity, AddHeartRate, AddCurrentTime, AddSpotifyStatus to handle specific tasks within BuildOSC.
3. Updated BuildOSC to call these new functions.
